requirements.txt
asgiref==3.7.2
Django==4.2.7
django-js-asset==2.1.0
django-mptt==0.15.0
sqlparse==0.4.4
